// src/models/pokemon.m.ts
export type Pokemon = {
    id: string;
    name: string;
    imggif: string;
    imglarge: string;
    imgnormal: string;
    type: Array<any>;
    type2: Array<any>;
    total: string;
    hp: string;
    attack: string;
    defense: string;
    sp_atk: string;
    sp_def: string;
    speed: string;
}